# [ratelimiter][perf] Use non-blocking rate limiter.