@ttsugriy 2023-06-22
you're very welcome, @antas-marcin! I love poking around weaviate codebase, so it's my pleasure :) Btw, there is one more change that's been sitting in review - https://github.com/weaviate/weaviate/pull/3148 It adds a missing test.